Category Crud Redesign

+++ public/js/Category.js
@@ -6,22 +6,8 @@ async function GetTableData() {
     ServerData = await res.json();
 }
 
-function CreateTable() {
-    let str = '';
-    for (let line of ServerData) {
-        str += `<tr>`;
-        str += `<td>${line.ID}</td>`;
-        str += `<td>${line.Category_Text}</td>`;
-        str += `<td onclick="DeleteLine(${line.ID})" class="tableBtn">מחק שורה</td>`;
-        str += `<td onClick="EditLine(${line.ID})" class="tableBtn">ערוך שורה</td>`;
-        str += `</tr>`;
-    }
-    document.querySelector('tbody').innerHTML = str;
-    document.getElementById('UpdateInput').value = "";
-}
-
 async function AddLine() {
-    let CatContent = document.getElementById('UpdateInput').value;
+    let CatContent = document.getElementById('Add-row-Input').value;
     if (CatContent === "") return;
     await fetch('/categories/Add', {
         method: 'POST',
@@ -32,6 +18,8 @@ async function AddLine() {
 }
 
 async function DeleteLine(id) {
+    let str = "Are you Sure you want to delete Category?"
+    if (!confirm(str)) return;
     await fetch('/categories/Del', {
         method: 'DELETE',
         headers: {'Content-Type': 'application/json'},
@@ -41,12 +29,42 @@ async function DeleteLine(id) {
 }
 
 async function EditLine(id) {
-    let CatContent = document.getElementById('UpdateInput').value;
-    if (CatContent === "") return;
+    let CatContent = document.getElementById('Update-row-Input').value;
+    if (CatContent === "") return alert("Please Enter Name");
     await fetch('/categories/Edit', {
         method: 'PATCH',
         headers: {'Content-Type': 'application/json'},
         body: JSON.stringify({'CatContent': CatContent, 'Cat_id': id})
     });
+    document.querySelector('.popup').close();
     GetTableData().then(CreateTable);
+}
+
+function CreateTable() {
+    let str = '';
+    for (let line of ServerData) {
+        str += `<tr>`;
+        str += `<td>${line.ID}</td>`;
+        str += `<td>${line.Category_Text}</td>`;
+        str += `<td onclick="DeleteLine(${line.ID})" class="tableBtn">מחק שורה</td>`;
+        str += `<td onClick="togglePopup(${line.ID},'${line.Category_Text}')" class="tableBtn">ערוך שורה</td>`;
+        str += `</tr>`;
+    }
+    document.querySelector('tbody').innerHTML = str;
+    document.getElementById('Add-row-Input').value = "";
+    document.getElementById('Update-row-Input').value = "";
+}
+
+function togglePopup(CatId = 0, CatName = " ") {
+    let popup = document.querySelector('.popup');
+    let span = document.querySelector('#Current-Category');
+    let btn = document.querySelector('#Update-Btn');
+    if (CatId) {
+        let str = `שם נוכחי: `;
+        str += `${CatName}.`;
+        span.innerHTML = str;
+        str = `<button type="submit" onclick="EditLine(${CatId})">ערוך שורה</button>`;
+        btn.innerHTML = str;
+        popup.showModal();
+    } else popup.close();
 }
